Adding READMEs, renaming temporary package name, updating tests to use simpler constructor
