+++ src/main/java/org/openstreetmap/atlas/geography/atlas/raw/slicing/changeset/ChangeSetHandler.java
@@ -10,7 +10,7 @@ import org.openstreetmap.atlas.geography.atlas.Atlas;
 import org.openstreetmap.atlas.geography.atlas.builder.AtlasSize;
 import org.openstreetmap.atlas.geography.atlas.items.Point;
 import org.openstreetmap.atlas.geography.atlas.packed.PackedAtlasBuilder;
-import org.openstreetmap.atlas.geography.atlas.raw.slicing.temporary.TemporaryLine;
+import org.openstreetmap.atlas.geography.atlas.raw.temporary.TemporaryLine;
 
 /**
  * Base class that contains common functionality for applying a given slicing change set to a raw
