+++ src/main/java/org/openstreetmap/atlas/geography/atlas/raw/temporary/TemporaryEntity.java
@@ -1,4 +1,4 @@
-package org.openstreetmap.atlas.geography.atlas.raw.slicing.temporary;
+package org.openstreetmap.atlas.geography.atlas.raw.temporary;
 
 import java.io.Serializable;
 import java.util.Map;
